chore: flatten the directory; remove the engine dir
